Fit and finish for macOS app 2.1.x
